requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
numpy==2.4.2
pandas==3.0.0
pyactiveresource==2.2.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
ShopifyAPI==12.7.0
six==1.17.0
urllib3==2.6.3
